[Validators] Refactor validators
Also fixed removal of deprecated function parameter at send_file in
flask
