+++ api/tools/validators.py
@@ -1,0 +1,111 @@
+import re
+import validators
+
+def is_valid_username(username):
+    """ User name cannot have double underscores, double dashes, nor double dots """
+    username_regex = re.compile(r'^(?!.*\.\..*)^(?!.*--.*)^(?!.*__.*)^[-a-zA-Z0-9_.]+$')
+    return username_regex.match(username)
+
+
+def split_addr(emailStr, encoding):
+    import re
+
+    regexStr = r'^([^@]+)@[^@]+$'
+    matchobj = re.search(regexStr, emailStr)
+    if not matchobj is None:
+        print("EMAIL ADDRESS " + matchobj.group(1))
+    else:
+        print("Did not match")
+        return None, None
+
+    return [matchobj.group(0), matchobj.group(1)]
+
+
+def sanitize_address(addr, encoding):
+    """
+    Format a pair of (name, address) or an email address string.
+    """
+    from email.utils import parseaddr
+    from email.errors import InvalidHeaderDefect, NonASCIILocalPartDefect
+    from email.header import Header
+    from email.headerregistry import Address
+
+    if not isinstance(addr, tuple):
+        addr = parseaddr(addr)
+    nm, addr = addr
+    localpart, domain = None, None
+    nm = Header(nm, encoding).encode()
+
+    try:
+        try:
+            addr.encode('ascii')
+        except UnicodeEncodeError:  # IDN or non-ascii in the local part
+            localpart, domain = split_addr(addr, encoding)
+
+        # An `email.headerregistry.Address` object is used since
+        # email.utils.formataddr() naively encodes the name as ascii (see #25986).
+        if localpart and domain:
+            address = Address(nm, username=localpart, domain=domain)
+            return str(address)
+
+        try:
+            address = Address(nm, addr_spec=addr)
+        except (InvalidHeaderDefect, NonASCIILocalPartDefect):
+            localpart, domain = split_addr(addr, encoding)
+            address = Address(nm, username=localpart, domain=domain)
+
+    except Exception as err:
+        print(" Address not valid " + str(err))
+        return None
+
+    return str(address)
+
+
+def check_email(email):
+    # Regex test
+    import re
+    if not email:
+        return False
+
+    # As it is, it will support more than one plus in the string, FIX is required
+    match = re.match('^[_a-z0-9-\+]+(\.[_a-z0-9-\+]+)*@[a-z0-9-]+(\.[a-z0-9-]+)*(\.[a-z]{2,4})$', email)
+    if match == None:
+        return False
+
+    return True
+
+
+def get_validated_email(email):
+    # https://stackoverflow.com/questions/8022530/how-to-check-for-valid-email-address
+
+    if not validators.email(email):
+        return get_response_error_formatted(400, {'error_msg': "Please provide a valid email"})
+
+    try:
+        if not email or len(email) == 0:
+            return get_response_error_formatted(400, {'error_msg': "Please provide a valid email"})
+
+        email_clean = sanitize_address(email, 'iso-8859-1')
+        if not email_clean or not check_email(email_clean):
+            return get_response_error_formatted(400,
+                                                {'error_msg': "Please provide a valid email " + email + " is no valid"})
+
+        print(" Email after sanitize_address " + str(email_clean))
+        return email_clean
+
+    except Exception as e:
+        return get_response_error_formatted(400,
+                                            {'error_msg': "Please provide a valid email " + email + " is no valid"})
+
+    return None
+
+
+def is_password_valid(password):
+    """ Check password policies
+        We should check a dictionary and length
+    """
+
+    if len(password) < 8:
+        return False
+
+    return True
